refactor(ui): align What's New with frontend style guides
- move content feed, validators and eligibility out of the slice into
  uiSrc/utils/whatsNew (slice keeps state only); eligibility tests move
  to utils/tests
- extract formatReleaseDate to WhatsNewModal.utils.ts and FeatureCard
  Props to FeatureCard.types.ts
- convert WhatsNewVersionType union to an enum and compare against it
- add whatsNewCardFactory/whatsNewVersionFactory (fishery + faker) and
  a factory-based FeatureCard spec
- tighten comments
