+++ redisinsight/ui/src/components/whats-new/components/feature-card/FeatureCard.types.ts
@@ -1,0 +1,6 @@
+import { WhatsNewCard } from 'uiSrc/constants/content/whatsNew.types'
+
+export interface Props {
+  card: WhatsNewCard
+  onLinkClick: (cardId: string, href: string) => void
+}
