+++ redisinsight/ui/src/slices/app/whatsNew.ts
@@ -2,71 +2,9 @@ import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 
 import { BrowserStorageItem } from 'uiSrc/constants'
 import { localStorageService } from 'uiSrc/services'
-import { isVersionHigher } from 'uiSrc/utils'
-import whatsNewContent from 'uiSrc/constants/content/whatsNew.json'
-import {
-  WhatsNewCard,
-  WhatsNewFeed,
-  WhatsNewVersion,
-} from 'uiSrc/constants/content/whatsNew.types'
+import { getLatestWhatsNewVersion } from 'uiSrc/utils'
 import { StateWhatsNew } from 'uiSrc/slices/interfaces'
 import { RootState } from 'uiSrc/slices/store'
-
-const isValidCard = (card: unknown): card is WhatsNewCard => {
-  const c = card as Partial<WhatsNewCard>
-  return (
-    typeof c?.id === 'string' &&
-    typeof c?.title === 'string' &&
-    typeof c?.body === 'string'
-  )
-}
-
-const isValidVersion = (entry: unknown): entry is WhatsNewVersion => {
-  const v = entry as Partial<WhatsNewVersion>
-  const isValid =
-    typeof v?.version === 'string' &&
-    Array.isArray(v?.cards) &&
-    v.cards.every(isValidCard)
-
-  if (!isValid) {
-    console.warn('[whatsNew] Skipping malformed content entry:', entry)
-  }
-  return isValid
-}
-
-/**
- * Bundled "What's new" content. Static import — no network fetch. Malformed
- * entries are dropped with a warning so one bad block can't break the modal.
- * Sorted by version descending so the latest release is always first.
- */
-export const whatsNewFeed: WhatsNewFeed = (
-  whatsNewContent.versions as unknown[]
-)
-  .filter(isValidVersion)
-  .sort((a, b) => (isVersionHigher(a.version, b.version) ? -1 : 1))
-
-export const getLatestWhatsNewVersion = (): WhatsNewVersion | undefined =>
-  whatsNewFeed[0]
-
-/**
- * Whether the modal should auto-open for `toVersion` after an update: the
- * version must exist in the feed, have cards, not be a patch release, and be
- * newer than the last version the user has already seen. Pure — the caller
- * supplies `lastVersionSeen` (from redux state or localStorage). Manual opens
- * never use this.
- */
-export const isWhatsNewEligible = (
-  toVersion: string,
-  lastVersionSeen: string | null,
-): boolean => {
-  const versionEntry = whatsNewFeed.find((v) => v.version === toVersion)
-  const hasCards = !!versionEntry?.cards?.length
-  const isAutoOpenable = versionEntry?.type !== 'patch'
-  const isNewerThanSeen =
-    !lastVersionSeen || isVersionHigher(toVersion, lastVersionSeen)
-
-  return !!versionEntry && hasCards && isAutoOpenable && isNewerThanSeen
-}
 
 export const initialState: StateWhatsNew = {
   isOpen: false,
@@ -85,9 +23,7 @@ const whatsNewSlice = createSlice({
       state.isOpen = true
       state.selectedVersion = payload ?? latest
 
-      // Opening the modal (auto or manual) means the user has now seen the
-      // latest release, so record it to prevent it auto-opening again. Always
-      // the latest version, never the browsed one.
+      // Any open counts as seeing the latest release — prevents re-auto-open.
       if (latest && latest !== state.lastVersionSeen) {
         state.lastVersionSeen = latest
         localStorageService.set(
